Move nmeareader and its redis connection to own file
